task2/requirements.txt
datasets
sentence-transformers
faiss-cpu
ragas
python-dotenv
openai
torch
transformers
gradio 